code refactoring, use modules and js minification

+++ src/Service/TemplateEngine.js
@@ -1,40 +1,52 @@
-const ejs = require('ejs');
-const fs = require('fs');
-const fse = require('fs-extra');
+import ejs from "ejs";
+import fs from "fs";
+import fse from "fs-extra";
 
-let dirFolder = './dist';
-let defaultOptions = {};
+const dirFolder = './dist';
+const defaultOptions = {};
 
-module.exports = {
-    serveMode: false,
-    async renderPage(page, data) {
-        return await ejs.renderFile('./views/layout.ejs', Object.assign({page: page}, data, {dev: this.serveMode}), defaultOptions);
-    },
-    createDistFolder() {
-        try {
-            fs.rmdirSync(dirFolder, { recursive: true });
-            fs.mkdirSync(dirFolder);
-        } catch(err) {
-            console.error(err);
-        }
-    },
-    viewExists(view) {
-        return fs.existsSync('./views/pages/'+ view +'.ejs')
-    },
-    saveFile(file, dir, content) {
-        fs.mkdirSync(dirFolder + dir, { recursive: true })
+const TemplateEngine = {}
 
-        try {
-            fs.writeFileSync(dirFolder + dir + "/" + file + ".html", content, {})
-        } catch(err) {
-            console.error(err);
-        }
-    },
-    copyAssets() {
-        let files = fs.readdirSync('./public')
+TemplateEngine.serveMode = false
 
-        for (const file of files) {
-            fse.copySync('./public/'+files, dirFolder + '/' + file);
-        }
+TemplateEngine.renderPage = async(page, data) => {
+    return await ejs.renderFile('./views/layout.ejs', Object.assign(
+        {page: page},
+        data,
+        {dev: TemplateEngine.serveMode}
+    ), defaultOptions);
+}
+
+TemplateEngine.createDistFolder = (serveMode) => {
+    try {
+        if (!serveMode) fs.rmdirSync(dirFolder, { recursive: true });
+
+        if (!fs.existsSync(dirFolder)) fs.mkdirSync(dirFolder);
+    } catch(err) {
+        console.error(err);
     }
 }
+
+TemplateEngine.viewExists = (view) => {
+    return fs.existsSync('./views/pages/'+ view +'.ejs')
+}
+
+TemplateEngine.saveFile = (file, dir, content) => {
+    fs.mkdirSync(dirFolder + dir, { recursive: true })
+
+    try {
+        fs.writeFileSync(dirFolder + dir + "/" + file + ".html", content, {})
+    } catch(err) {
+        console.error(err);
+    }
+}
+
+TemplateEngine.copyAssets = () => {
+    let files = fs.readdirSync('./public')
+
+    for (const file of files) {
+        fse.copySync('./public/'+files, dirFolder + '/' + file);
+    }
+}
+
+export default TemplateEngine;
